requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.23
Flask-Cors==4.0.0
PyMySQL==1.1.1
python-dotenv==1.0.1
# Optional (recommended)
alembic==1.13.1
marshmallow==3.21.1
PyJWT==2.8.0
Flask-Mail
Flask-Limiter
APScheduler
flask_migrate 

